// generated/src/app/api-models/special-programs.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class SpecialPrograms {
  specialProgramId: string;
  specialProgramDesc: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}